# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "MCN service API",
    "description": "MCN service api - all public facing methods",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "basePath": "/api",
  "paths": {
    "/users": {
      "get": {
        "x-swagger-router-controller": "api",
        "operationId": "getUser",
        "tags": ["/users"],
        "description": "get user data from the database.",
        "parameters": [
          {
           "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          }
        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/UserCreated"
            }
          }
        }
      },
      "post": {
        "x-swagger-router-controller": "api",
        "operationId": "createUser",
        "tags": ["/users"],
        "description": "create user in database.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          }
        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/UserCreated"
            }
          }
        }
      }


    },


    "/match": {

      "get": {
        "x-swagger-router-controller": "api",
        "operationId": "getMatchForUser",
        "tags": ["/match"],
        "description": "find potential matches for user",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          }

        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/matchResponse"
            }
          }
        }
      }


    },



    "/connection": {
      "post": {
        "x-swagger-router-controller": "api",
        "operationId": "connectWithUser",
        "tags": [
          "/connect"
        ],
        "description": "connect with a specific user ",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          },
          {
            "name": "fb_target_id",
            "in": "query",
            "description": "The FB user id of the target user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/connectionResponse"
            }
          }
        }
      },
      "put": {
        "x-swagger-router-controller": "api",
        "operationId": "updateConnectionStatusWithUser",
        "tags": [
          "/connect"
        ],
        "description": "accept/reject connection request ",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          },
          {
            "name": "fb_target_id",
            "in": "query",
            "description": "The FB user id of the target user",
            "required": true,
            "type": "string"
          },
          {
            "name": "connectionStatus",
            "in": "body",
            "description": "status to update for the connection - reject or accept",
            "required": true,
            "schema": {
              "$ref": "#/definitions/connectionStatus"
            }
          }

        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request",
            "schema": {
              "$ref": "#/definitions/connectionResponse"
            }
          }
        }
      }
    },





    "/users/preferences":{

      "get": {
        "x-swagger-router-controller": "api",
        "operationId": "getUserPref",
        "tags": ["/users"],
        "description": "get user preferences from the database.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          }
        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request.",
            "schema": {
              "$ref": "#/definitions/updatePref"
            }
          }
        }
      },
      "put": {
        "x-swagger-router-controller": "api",
        "operationId": "updateUserPref",
        "tags": ["/users"],
        "description": "update user pref in database.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "required": true,
            "description": "The authorization token",
            "default": "Bearer"
          },
          {
            "name": "preferences",
            "in": "body",
            "description": "user preferences",
            "required": true,
            "schema": {
              "$ref": "#/definitions/updatePref"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "200": {
            "description": "Successful request."
          }
        }
      }

    }

  },
  "definitions": {


    "UserCreated":{
      "type": "object",
      "properties": {
        "userID":{
          "type": "string"
        }
      },
      "required" : ["userID"]

    },



    "connectionResponse":{
      "type": "string"

    },


    "matchResponse":{
      "type": "array",
      "items": {
        "$ref": "#/definitions/UserCreated"
      }
    },

    "connectionStatus":{
      "type" : "object",
      "required": [
        "status"
      ],
      "properties":{
        "status":{
          "type":"string",
          "enum": ["accepted", "rejected", "pending"]
        }
      }
    },



    "updatePref":{
      "type" : "object",
      "required": [
        "currency",
        "operation",
        "amount"
      ],
      "properties":{
        "currency":{
          "type":"string"
        },
        "operation":{
          "type": "string"
        },
        "amount":{
          "type": "integer"
        },
        "publicProfile":{
          "type": "boolean"
        }

      }

    },


    "Error": {
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      },
      "required": ["errorMessage"]
    }



  }
}